requirements.txt
lightning==2.5.0.post0
spconv-cu120==2.3.6
trimesh>=4.6.4
addict>=2.4.0
scipy>=1.15.2
h5py>=3.13.0
tqdm>=4.67.1
torchmetrics>=1.6.3
hydra-core>=1.3.2
wandb==0.20.1
matplotlib
rich